requirements.txt
django
psycopg2-binary
django-geosource
django-leaflet
django-geojson
djangorestframework
geopy